requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.1.2
pillow==10.2.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2024.1
reportlab==4.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
